Module Linked_set

module Make (Atomic : sig ... end) (Size : sig ... end) : sig ... end

Functorized lock-free linked set with length for testing and as an example of using Size. The functorization is to allow the use of traced atomics with DSCheck.