Module Saturn_lockfree.Queue_unsafe

type 'a t

The type of lock-free queue.

val create : unit -> 'a t

create () returns a new queue, initially empty.

val is_empty : 'a t -> bool

is_empty q returns empty if q is empty.

val push : 'a t -> 'a -> unit

push q v adds the element v at the end of the queue q.

exception Empty

Raised when pop or peek is applied to an empty queue.

val pop_exn : 'a t -> 'a

pop q removes and returns the first element in queue q.

raises Empty
if q is empty.

val pop_opt : 'a t -> 'a option

pop_opt q removes and returns the first element in queue q, or returns None if the queue is empty.

val peek_exn : 'a t -> 'a

peek q returns the first element in queue q.

raises Empty
if q is empty.

val peek_opt : 'a t -> 'a option

peek_opt q returns the first element in queue q, or returns None if the queue is empty.